requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# HTTP client for proxy
httpx>=0.26.0

# Templates
jinja2>=3.1.0

# Environment
python-dotenv>=1.0.0

# Encryption for provider keys
cryptography>=42.0.0

# Token counting
tiktoken>=0.5.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
